+++ packages/xlsx-preview/src/api-worksheet.ts
@@ -19,6 +19,7 @@ import {
   ImageCollection,
   PivotCollection,
   MergeCollection,
+  ShapeCollection,
   SparklineGroupCollection,
   TableCollection,
   ThreadedNotesCollection,
@@ -81,6 +82,7 @@ export class Worksheet {
   readonly tables: TableCollection;
   readonly charts: ChartCollection;
   readonly images: ImageCollection;
+  readonly shapes: ShapeCollection;
   readonly sparklineGroups: SparklineGroupCollection;
   readonly pivots: PivotCollection;
   readonly freeze: SheetFreeze;
@@ -102,6 +104,7 @@ export class Worksheet {
     this.tables = new TableCollection(handle, this.sheetRef);
     this.charts = new ChartCollection(handle, this.sheetRef);
     this.images = new ImageCollection(handle, this.sheetRef);
+    this.shapes = new ShapeCollection(handle, this.sheetRef);
     this.sparklineGroups = new SparklineGroupCollection(handle, this.sheetRef);
     this.pivots = new PivotCollection(handle, this.sheetRef);
     this.freeze = new SheetFreeze(handle, this.sheetRef);